pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatch-vcs>=0.4",
  "hatchling>=1.25",
]

[project]
name = "django-admin-sync"
description = ""
readme = "README.md"
authors = [
  { name = "sax", email = "s.apostolico@gmail.com" },
]
requires-python = ">=3.12"
classifiers = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

dynamic = [
  "version",
]

dependencies = [
  "django-admin-extra-buttons>=2.1",
  "django-concurrency",
  "django-stubs",
  "natural-keys>=2.1.1",
  "pytz>=2025.1",
  "requests",
]

[dependency-groups]
dev = [
  "django-constance>=4.3.2",
  #  "check-manifest",
  #  "coverage>4.4",
  #  "django-coverage-plugin>=3.1.0",
  "django-factory-boy",
  "django-webtest>=1.9.13",
  "freezegun",
  "ipdb>=0.13.13",
  "pre-commit",
  "psycopg2-binary",
  "pytest",
  "pytest-cov",
  "pytest-django",
  "pytest-echo",
  "pytest-responses",
  "tox",
]

docs = [
  "markdown-include>=0.8.1",
  "mdx-gh-links>=0.4",
  "mkdocs>=1.6.1",
  "mkdocs-autolinks-plugin>=0.7.1",
  "mkdocs-autorefs>=1.3.1",
  "mkdocs-awesome-pages-plugin>=2.10.1",
  "mkdocs-ezlinks-plugin>=0.1.14",
  "mkdocs-glightbox>=0.4",
  "mkdocs-include-markdown-plugin>=6.2.2",
  "mkdocs-macros-plugin>=1.3.7",
  "mkdocs-markdownextradata-plugin>=0.2.6",
  "mkdocs-material>=9.6.5",
  "mkdocs-material-extensions>=1.3.1",
  "mkdocstrings>=0.28.1",
  "mkdocstrings-python>=1.16.1",
]

mypy = [
  "django-stubs",
  "mypy",
  "types-requests",
]

[tool.hatch]
build.hooks.vcs.version-file = "src/admin_sync/version.py"
version.source = "vcs"
raw-options = { root = "../.." }

[tool.hatch.build.targets.sdist]
ignore-vcs = true
include = [
  "src/admin_sync/**/*.*",
]
exclude = [
  "~*.*",
]

[tool.hatch.build.targets.wheel]
packages = [ "src/admin_sync" ]
exclude = [
  "~*.*",
]

[tool.pyproject-fmt]
max_supported_python = "3.13"

[tool.uv]
package = true
